@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <UserSecretsId>2480691f-5bcb-4256-9d55-508a7f38a01b</UserSecretsId>
     <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
 
@@ -17,16 +17,25 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
+    <!--<PackageReference Include="Microsoft.AspNetCore.App" />-->
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.0.0">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="5.6.0" />
  
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.2.0" />
-    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="6.0.0" />   
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
-    <PackageReference Include="System.Linq.Dynamic.Core" Version="1.0.8.8" />
+    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.0.0" />
+    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="6.0.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />   
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc4" />   
+    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="5.6.0" />
+    <PackageReference Include="System.Linq.Dynamic.Core" Version="1.0.19" />
   </ItemGroup>
 
-  <ItemGroup> 
-    <ProjectReference Include="..\WebAdmin.Dal\WebAdmin.DAL.csproj" />
+  <ItemGroup>
+    <ProjectReference Include="..\WebAdmin.EF\WebAdmin.EF.csproj" />
     <ProjectReference Include="..\WebAdmin.Interface\WebAdmin.Interfaces.csproj" />
     <ProjectReference Include="..\WebAdmin.Entities\WebAdmin.Entities.csproj" />
     <ProjectReference Include="..\WebAdmin.ViewModels\WebAdmin.ViewModels.csproj" />
