@@ -1,3 +1,3 @@
-@using Core3Ang8
-@namespace Core3Ang8.Pages
+@using AdminUI
+@namespace AdminUI.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
