@@ -4,7 +4,7 @@
 
 <div class="container">
     <div class="jumbotron">
-        <h1>DOT&amp;NET GYM</h1>      
+        <h1>DOT&amp;NET Core Admin WebAPI</h1>      
         <p>All APIS Service of Project</p>
     </div>  
 </div>
