@@ -26,7 +26,7 @@
                     <span class="icon-bar"></span>
                     <span class="icon-bar"></span>
                 </button>
-                <a asp-area="" asp-controller="Home" asp-action="Index" class="navbar-brand">DOT&amp;NET GYM</a>
+                <a asp-area="" asp-controller="Home" asp-action="Index" class="navbar-brand">DOT&amp;NET WebAPI</a>
             </div>
            
         </div>
